# api/dashboard.py
"""调度看板 + 运维概览 API（参考 codex2api SchedulerBoard / ops.go）。"""

from __future__ import annotations

import asyncio
import datetime
import json
import os
import platform
import shutil
import time

from fastapi import APIRouter, Header
from fastapi.concurrency import run_in_threadpool
from fastapi.responses import PlainTextResponse, StreamingResponse

from api.support import require_identity
from services.account_service import AccountService, account_service
from services.config import config, DATA_DIR
from services.image_service import storage_stats
from services.log_service import log_service
from services.metrics_service import metrics_service


def _collect_account_health(accounts: list[dict]) -> dict[str, object]:
    """按健康档位统计账号分布。"""
    tiers = {"healthy": 0, "warm": 0, "risky": 0, "banned": 0}
    statuses: dict[str, int] = {}
    total_quota = 0
    total_inflight = 0
    for account in accounts:
        status = str(account.get("status") or "未知")
        statuses[status] = statuses.get(status, 0) + 1
        total_quota += max(0, int(account.get("quota") or 0))
        total_inflight += int(account.get("image_inflight") or 0)
        if status in {"禁用", "异常"}:
            tiers["banned"] += 1
        else:
            tier = AccountService._account_health_tier(account)
            tiers[tier] = tiers.get(tier, 0) + 1
    return {
        "tiers": tiers,
        "statuses": statuses,
        "total": len(accounts),
        "total_quota": total_quota,
        "total_inflight": total_inflight,
    }


def _collect_log_stats() -> dict[str, object]:
    """基于系统日志统计用量（近 24h 成功/失败/总量）。"""
    logs = log_service.list(limit=1000)
    now = time.time()
    day_ago = now - 86400
    success = 0
    failed = 0
    calls: dict[str, int] = {}
    recent: list[dict[str, object]] = []
    for item in logs:
        created = item.get("created_at") or ""
        try:
            ts = time.mktime(datetime.datetime.strptime(created[:19], "%Y-%m-%d %H:%M:%S").timetuple())
        except (ValueError, TypeError):
            ts = 0
        if ts < day_ago:
            continue
        status = str(item.get("status") or "success")
        summary = str(item.get("summary") or "调用")
        calls[summary] = calls.get(summary, 0) + 1
        if status == "failed":
            failed += 1
        else:
            success += 1
        recent.append({"time": created, "summary": summary, "status": status})
    return {
        "success_24h": success,
        "failed_24h": failed,
        "total_24h": success + failed,
        "by_summary": calls,
        "recent": recent[-20:],
    }


def _collect_ops_overview() -> dict[str, object]:
    """运维概览：CPU/内存/磁盘/进程/存储统计。"""
    # CPU 使用率（Windows 上取不到精确值，用 loadavg 兜底）
    cpu_percent: float | None = None
    try:
        import psutil  # type: ignore

        cpu_percent = psutil.cpu_percent(interval=0.1)
        memory = psutil.virtual_memory()
        memory_used = memory.used
        memory_total = memory.total
    except ImportError:
        memory_used = memory_total = 0
    if cpu_percent is None:
        try:
            cpu_percent = os.getloadavg()[0] / max(1, os.cpu_count() or 1) * 100
        except (OSError, AttributeError):
            cpu_percent = 0.0
    disk = shutil.disk_usage(str(DATA_DIR))
    return {
        "platform": platform.platform(),
        "python": platform.python_version(),
        "pid": os.getpid(),
        "uptime_seconds": int(time.time() - _PROCESS_START_TIME),
        "cpu_percent": round(cpu_percent, 1),
        "memory_used_mb": round(memory_used / 1024 / 1024, 1) if memory_used else None,
        "memory_total_mb": round(memory_total / 1024 / 1024, 1) if memory_total else None,
        "disk_free_mb": round(disk.free / 1024 / 1024, 1),
        "disk_total_mb": round(disk.total / 1024 / 1024, 1),
        "storage": storage_stats(),
        "scheduler_mode": config.scheduler_mode,
        "refresh_account_interval_minute": config.refresh_account_interval_minute,
        "image_account_concurrency": config.image_account_concurrency,
    }


_PROCESS_START_TIME = time.time()


def create_router() -> APIRouter:
    router = APIRouter()

    @router.get("/api/dashboard/scheduler")
    async def scheduler_dashboard(authorization: str | None = Header(default=None)):
        """调度看板：账号健康分布 + 实时并发 + 调度分排名。"""
        require_identity(authorization)
        accounts = account_service.list_accounts()
        health = _collect_account_health(accounts)
        # 带调度分的账号排名（供前端展示）
        ranked = []
        for account in accounts:
            if account.get("status") in {"禁用", "异常"}:
                continue
            token = str(account.get("access_token") or "")
            tier = AccountService._account_health_tier(account)
            score = AccountService._account_dispatch_score(account, tier)
            ranked.append(
                {
                    "email": account.get("email"),
                    "type": account.get("type"),
                    "status": account.get("status"),
                    "quota": max(0, int(account.get("quota") or 0)),
                    "success": int(account.get("success") or 0),
                    "fail": int(account.get("fail") or 0),
                    "image_inflight": int(account.get("image_inflight") or 0),
                    "tier": tier,
                    "score": score,
                    "priority": account_service._priority_for_token(token),
                }
            )
        ranked.sort(key=lambda item: (item["tier"] != "healthy", -item["score"]))
        return {"health": health, "accounts": ranked}

    @router.get("/api/dashboard/ops")
    async def ops_overview(authorization: str | None = Header(default=None)):
        """运维概览：CPU/内存/磁盘/账号池/日志统计。"""
        require_identity(authorization)
        return await run_in_threadpool(_collect_ops_overview)

    @router.get("/api/dashboard/usage")
    async def usage_stats(authorization: str | None = Header(default=None)):
        """用量统计：近 24h 调用量 + 按类型分布 + 最近记录。"""
        require_identity(authorization)
        return await run_in_threadpool(_collect_log_stats)

    @router.get("/api/dashboard/latency")
    async def latency_stats(authorization: str | None = Header(default=None)):
        """请求延迟统计：总请求/错误率/平均延迟/按路径分布/在途。"""
        require_identity(authorization)
        return metrics_service.get_summary()

    @router.get("/metrics", include_in_schema=False)
    async def prometheus_metrics():
        """Prometheus 指标端点（无需鉴权，供监控系统抓取）。"""
        return PlainTextResponse(metrics_service.prometheus(), media_type="text/plain; version=0.0.4")

    @router.get("/api/dashboard/stream", include_in_schema=False)
    async def dashboard_stream(authorization: str | None = Header(default=None), token: str = ""):
        """SSE 实时推送：看板数据每 3 秒推送一次。

        EventSource 无法传 Authorization header，支持 ?token= 查询参数鉴权。
        """
        if not authorization and token:
            authorization = f"Bearer {token}"
        require_identity(authorization)

        async def event_generator():
            try:
                while True:
                    try:
                        accounts = account_service.list_accounts()
                        health = _collect_account_health(accounts)
                        payload = {
                            "type": "dashboard",
                            "ts": int(time.time()),
                            "health": health,
                            "latency": metrics_service.get_summary(),
                        }
                        yield f"data: {json.dumps(payload, ensure_ascii=False)}\n\n"
                    except Exception:
                        # 单次数据构建失败，跳过本次，等待下一周期
                        pass
                    await asyncio.sleep(3)
            except asyncio.CancelledError:
                # 客户端断开连接，正常退出协程
                return

        return StreamingResponse(
            event_generator(),
            media_type="text/event-stream",
            headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
        )

    return router