# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ChatGPT2API
  description: |
    OpenAI 兼容的图片生成/编辑 API 代理。所有接口与 OpenAI 协议兼容。
    鉴权方式：HTTP Bearer Token（Authorization: Bearer <auth-key>）。
  version: 1.9.0
servers:
  - url: http://localhost:23456
    description: 本地开发
security:
  - bearerAuth: []
paths:
...
  /api/dashboard/latency:
    get:
      summary: 请求延迟统计
      tags: [Observability]
      responses:
        '200':
          description: 延迟统计摘要
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_requests:
                    type: integer
                  total_errors:
                    type: integer
                  avg_latency_ms:
                    type: number
                  error_rate:
                    type: number
                  by_path:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            param:
              type: string
              nullable: true
            code:
              type: string
  responses:
    Unauthorized:
      description: 鉴权失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: 密钥无效或已失效，请重新登录
              type: authentication_error
              code: invalid_api_key
...